% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fuzzy_intersect}
\alias{fuzzy_intersect}
\title{Fuzzy intersection of item loadings on personality domains}
\usage{
fuzzy_intersect(...)
}
\arguments{
\item{...}{Vectors of loadings labelled with item names}
}
\description{
Given a set of items loadings, calculate the minimum
loading for each item on all the dimensions.
}

